requirements.txt
# Core dependencies
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
Pillow>=10.0.0
opencv-python>=4.8.0
moviepy>=1.0.3
networkx>=3.0.0
markdown>=3.5.0
qrcode>=7.4.2

# Web frameworks
streamlit>=1.26.0
Flask==2.2.5
flask-cors==4.0.0

# Music processing
librosa>=0.10.0
opensmile>=1.0.0
youtube-dl>=2021.12.17
yt-dlp>=2023.7.6

# AI and ML
openai>=1.0.0
replicate>=0.17.0
faiss-cpu>=1.7.0
ollama>=0.1.6
openai-agents
#scikit-learn>=1.0.0
#transformers>=4.36.0

# Web scraping and API
requests>=2.31.0
beautifulsoup4>=4.12.0
wikipedia-api>=0.6.0
spotipy>=2.23.0
duckduckgo-search>=3.9.0
wikipedia>=1.4.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0
customtkinter>=5.2.0
colorama>=0.4.6